analytics service and test moved to analytic folder
