+++ analytics/analytics-ui/src/app/parameters/parameter-form/parameter-form.component.ts
@@ -15,7 +15,7 @@
  */
 
 import { Component, OnInit, Input, ViewChild } from '@angular/core';
-import { AnalyticsService } from '../../services/analytics.service';
+import { AnalyticsService } from '../../analytics/analytics.service';
 
 @Component({
   selector: 'app-parameter-form',
